[FAB-7819] Fix the CI failure
CI fails without reporting any error.

After tracking the problem, the root cause is that in the new added
chaincode_install test, the pem file is not read as binary, which caused
grpc channel creation problem and quit.

This patchset fixes the problem and regenerates the protobuf files with
python 3.6.

And refine the code structure and minor update the doc.

TODO: we need to enable the chaincode installation testing with an env
that has channel created and joined already.

Change-Id: I7ff5939b6b13c15e3ccdadd2d2804f8b17cf369c
Signed-off-by: Baohua Yang <yangbaohua@gmail.com>